requirements.txt
PyYAML~=6.0.1
openai~=1.10.0
torch~=2.3.0
transformers~=4.36.0
fastapi~=0.109.1
uvicorn~=0.25.0
pydantic~=1.10.12
requests~=2.31.0
bson~=0.5.10
pymongo~=4.6.0
PyGithub~=2.1.1
chainlit~=1.0.200
tavily-python